+++ changelog.md
@@ -2,7 +2,7 @@
 
 #### v0.2.0
 
-* 
+* Database/Db.php - throw an execption when function does not exist
 
 
 #### v0.1.0